feat: Add `vaccel_exec` for ORB-SLAM2 track func
Introduce a `vaccel_exec` call to invoke the track_stereo function of
the GPU-accelerated clustering ORB-SLAM2 implementation.

By default, the `VACCEL` CMake option is set to `OFF`. The `build.sh`
script enables this option to build the liborb.so shared library and
generate the `build/Test` executable, which uses vaccel_exec to run the
tracking logic.
Signed-off-by: Maria Gkeka <mgkeka@nubificus.co.uk>